matching-api/pyproject.toml
[project]
name = "matching-api"
version = "0.1.0"
description = "REST API for module matching and recognition recommendations"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "chromadb>=0.4",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "reportlab>=4.4.5",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
]

[project.scripts]
load-units = "scripts.load_units:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "httpx>=0.27",
]
